Refactor: Consolidate MatchStatistics definition and update imports for improved organization
